Add database tests
